1. Put Request 2. Gson library replaced by Jackson 3. Added libraries
1. Added a PUT method to the "localhost:9000/{threadid}/{postid}" URL
pattern
2. Changed the project to use the Jackson libraries and Restlet
extension for Object-Json serialization & deserialization instead of the
Gson one. Firstly it's required by the restlet Swagger extension, and it
doesn't make sense to have two libaries which exist to carry out the
same function. Secondly it makes the code easier to read and write with
fewer method calls.
3. I added some libraries to the root directory, and more libraries in
the lib directory which are needed by the jackson jar. All libraries
should be added to the classpath. As far as I know the order in which
this is done shouldn't matter.